# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Shop Liên Quân API - OpenAPI Spec",
    "version": "1.0.0",
    "description": "Tài liệu OpenAPI 3.0 / JSON API đầy đủ cho hệ thống website bán tài khoản game Liên Quân Mobile.",
    "contact": {
      "name": "API Support",
      "email": "support@shoplienquan.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT Access Token dạng: Bearer <token>"
      }
    },
...
  },
  "paths": {
...
    "/api/transactions": {
      "get": {
        "summary": "Lấy lịch sử giao dịch của tôi",
        "tags": [
          "Transaction"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công"
          }
        }
      }
    },
...
    "/api/admin/dashboard": {
      "get": {
        "summary": "Thống kê dashboard admin",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy dữ liệu dashboard thành công"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền admin"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "summary": "Danh sách người dùng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách người dùng thành công"
          },
          "401": {
            "description": "Chưa đăng nhập"
          },
          "403": {
            "description": "Không có quyền admin"
          }
        }
      }
    },
    "/api/admin/users/{id}": {
      "put": {
        "summary": "Cập nhật người dùng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/users/{id}/money": {
      "post": {
        "summary": "Cộng hoặc trừ tiền người dùng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16
            },
            "description": "Khóa duy nhất cho một lần điều chỉnh số dư; giữ nguyên khi retry."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "amount"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "add",
                      "sub"
                    ]
                  },
                  "amount": {
                    "type": "integer",
                    "example": 100000
                  },
                  "description": {
                    "type": "string",
                    "example": "Admin cộng tiền test"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật số dư thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "404": {
            "description": "Không tìm thấy người dùng"
          }
        }
      }
    },
    "/api/admin/accounts": {
      "get": {
        "summary": "Danh sách tài khoản game",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/orders": {
      "get": {
        "summary": "Danh sách đơn hàng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/transactions": {
      "get": {
        "summary": "Danh sách giao dịch",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/sales": {
      "get": {
        "summary": "Danh sách sale",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Thêm sale",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/sales/{id}": {
      "put": {
        "summary": "Cập nhật sale",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Tắt sale",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/discounts": {
      "get": {
        "summary": "Danh sách mã giảm giá",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Thêm mã giảm giá",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/discounts/{id}": {
      "put": {
        "summary": "Cập nhật mã giảm giá",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "summary": "Tắt mã giảm giá",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/setting": {
      "get": {
        "summary": "Lấy cấu hình website",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy cấu hình website thành công"
          }
        }
      },
      "put": {
        "summary": "Cập nhật cấu hình website",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cập nhật cấu hình website thành công"
          }
        }
      }
    },
    "/api/admin/logs": {
      "get": {
        "summary": "Lịch sử hoạt động hệ thống",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy lịch sử hoạt động thành công"
          }
        }
      }
    },
    "/api/admin/banks": {
      "get": {
        "summary": "Danh sách ngân hàng admin quản lý",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách thành công"
          }
        }
      },
      "post": {
        "summary": "Thêm ngân hàng nạp tiền mới",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bank_name",
                  "account_number",
                  "account_holder"
                ],
                "properties": {
                  "bank_name": {
                    "type": "string",
                    "example": "MBBank"
                  },
                  "account_number": {
                    "type": "string",
                    "example": "0987654321"
                  },
                  "account_holder": {
                    "type": "string",
                    "example": "NGUYEN VAN A"
                  },
                  "qr_template": {
                    "type": "string",
                    "example": "https://img.vietqr.io/image/MB-0987654321-compact2.png"
                  },
                  "status": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thêm thành công"
          }
        }
      }
    },
    "/api/admin/banks/{id}": {
      "put": {
        "summary": "Cập nhật thông tin ngân hàng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cập nhật thành công"
          }
        }
      },
      "delete": {
        "summary": "Xóa / ẩn ngân hàng",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa thành công"
          }
        }
      }
    },
...
    "/api/accounts": {
      "get": {
        "summary": "Lấy danh sách tài khoản đang bán",
        "tags": [
          "Account"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "loai_id",
            "schema": {
              "type": "integer"
            },
            "description": "Lọc theo loại tài khoản"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "price_asc",
                "price_desc"
              ]
            },
            "description": "Sắp xếp theo giá"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 20
            }
          }
        ]
      },
      "post": {
        "summary": "Thêm tài khoản game",
        "tags": [
          "Account"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "loai_id",
                  "gia",
                  "login"
                ],
                "properties": {
                  "loai_id": {
                    "type": "integer"
                  },
                  "thong_tin": {
                    "type": "string"
                  },
                  "list_thong_tin": {
                    "type": "string"
                  },
                  "img": {
                    "type": "string"
                  },
                  "list_img": {
                    "type": "string"
                  },
                  "login": {
                    "type": "string"
                  },
                  "gia": {
                    "type": "integer"
                  },
                  "ck": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}